% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{ch.row}
\alias{ch.row}
\title{Returns vector of checkmarks}
\usage{
ch.row(name, yesno)
}
\arguments{
\item{name}{name of row}

\item{yesno}{logical for checkmark}
}
\description{
Returns vector of checkmarks
}
\examples{
ch.row()

}
