% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conley.R
\name{ConleySEs}
\alias{ConleySEs}
\title{Returns conley SEs corrected for spatial autocorrelation}
\usage{
ConleySEs(reg, unit, time, lat, lon, kernel = "bartlett",
  dist_fn = "Haversine", dist_cutoff = 500, lag_cutoff = 5,
  lat_scale = 111, verbose = FALSE, cores = 1,
  balanced_pnl = FALSE)
}
\arguments{
\item{reg}{a regression-like object}

\item{unit}{character indicating unit}

\item{time}{character indicating time}

\item{lat}{character indicating latitude}

\item{lon}{character indicating longitude}

\item{kernel}{kernel for weights. defualts to bartlett}

\item{dist_fn}{distance function. defaults to haversine}

\item{dist_cutoff}{distance for cutoff. defaults to 500}

\item{lag_cutoff}{lag of cutoff. defaults to 5}

\item{lat_scale}{scale. defaults to 111}

\item{verbose}{verbose output. defaults to FALSE}

\item{cores}{number of cores to use. defaults to 1}

\item{balanced_pnl}{logical for whether balanced panel. defaults to FALSE}
}
\description{
Returns conley SEs corrected for spatial autocorrelation
}
\examples{
ConleySEs()

}
