% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\docType{data}
\name{font_legend_stuff}
\alias{font_legend_stuff}
\title{Shom's Custom ggplot2 themes}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 5.}
\usage{
font_legend_stuff
}
\value{
None
}
\description{
Custom ggplot legend for pretty plots
}
\keyword{datasets}
