% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{format_and_save_tab}
\alias{format_and_save_tab}
\title{Returns cluster robust first-stage f-stat from felm object}
\usage{
format_and_save_tab(table, path)
}
\arguments{
\item{table}{a latex table from stargazer}

\item{path}{a filepath to save the table}
}
\description{
Returns cluster robust first-stage f-stat from felm object
}
\examples{
format_and_save_tab()

}
