% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{get_fstat_felm}
\alias{get_fstat_felm}
\title{Returns cluster robust first-stage f-stat from felm object}
\usage{
get_fstat_felm(x)
}
\arguments{
\item{x}{An felm model}
}
\value{
fstat
}
\description{
Returns cluster robust first-stage f-stat from felm object
}
\examples{
get_fstat_felm()

}
