% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{got_Internet}
\alias{got_Internet}
\title{Checks to see if you have an internet connection}
\usage{
got_Internet()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
Checks to see if you have an internet connection
}
\examples{
got_Internet()

}
