% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{plot_rdd_binned}
\alias{plot_rdd_binned}
\title{Shom's Custom ggplot2 themes}
\usage{
plot_rdd_binned(data, forcing, outcome, cutoff = 0, weights = NULL,
  ylab = "Outcome", xlab = "Running Variable", bw = NULL, se = T)
}
\arguments{
\item{data}{A dataframe}

\item{forcing}{The forcing/running variable as a string}

\item{outcome}{The outcome varaible as a string}

\item{cutoff}{The cutoff. Defaults to 0.}

\item{weights}{Any weights you want to use as a string. Defaults to NULL.}

\item{ylab}{The label for the y axis.}

\item{xlab}{The label for the x axis.}

\item{bw}{The bandwidth. Defaults to the entire support of the forcing variable.}

\item{se}{Logical for whether to include the standard errors in the plot.}
}
\value{
a list of the plot (in ggplot) and the corresponding binned data
}
\description{
Custom ggplot theme for choropleths
}
