% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{rdd_robust_bw}
\alias{rdd_robust_bw}
\title{Robustness to bandwidth choice}
\usage{
rdd_robust_bw(data, forcing, outcome, cutoff = 0, weights = NULL,
  ylab = "Outcome", xlab = "Running Variable", bw = NULL, se = T)
}
\arguments{
\item{data}{A dataframe}

\item{forcing}{The forcing/running variable as a string}

\item{outcome}{The outcome varaible as a string}

\item{cutoff}{The cutoff. Defaults to 0.}

\item{weights}{Any weights you want to use as a string. Defaults to NULL.}

\item{ylab}{The label for the y axis.}

\item{xlab}{The label for the x axis.}

\item{bw}{The bandwidth. Defaults to half the IK bw to twice the IK bw.}

\item{se}{Logical for whether to include the standard errors in the plot.}
}
\value{
a list of the plot (in ggplot) and the corresponding effect estimates and standard errors
}
\description{
Robustness to bandwidth choice
}
