% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{stata_codebook}
\alias{stata_codebook}
\title{Makes a stata codebook}
\usage{
stata_codebook(x)
}
\arguments{
\item{x}{Takes in a vector of labelled values}
}
\value{
None
}
\description{
Makes a stata codebook
}
