% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{theme_shom}
\alias{theme_shom}
\title{Shom's Custom ggplot2 themes}
\usage{
theme_shom(family = c("Minion Pro", "Palatino"))
}
\arguments{
\item{family}{font family. defaults to Minion Pro}
}
\value{
None
}
\description{
Custom ggplot theme for general plots
}
