% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{theme_shom_pretty}
\alias{theme_shom_pretty}
\title{Shom's Custom ggplot2 themes}
\usage{
theme_shom_pretty(base_size = 11, waffle = FALSE)
}
\arguments{
\item{base_size}{base size font}

\item{waffle}{logical for waffle plot}
}
\value{
None
}
\description{
Custom ggplot theme for pretty plots
}
