% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{get_rse}
\alias{get_rse}
\title{Checks to see if you have an internet connection}
\usage{
get_rse(models.list, hc = "HC2")
}
\arguments{
\item{models.list}{A list of regression models (lm)}

\item{hc}{Type of robust standard error. Defaults to HC2.}
}
\value{
se.list
}
\description{
Checks to see if you have an internet connection
}
\examples{
get_rse()

}
